added a little changes in design

+++ server/Films/controller.js
@@ -86,7 +86,9 @@ const editFilm = async (req, res) => {
 const deleteFilm = async (req, res) => {
   const film = await Film.findById(req.params.id);
   if (film) {
-    fs.unlinkSync(path.join(__dirname + "../../../public/" + film.image));
+    if (fs.existsSync(path.join(__dirname + "../../../public/" + film.image))) {
+      fs.unlinkSync(path.join(__dirname + "../../../public/" + film.image));
+    }
     await Film.deleteOne({ _id: req.params.id });
     res.status(200).send("ok");
   } else {
